Update project v1.1.1

+++ models/company.go
@@ -11,7 +11,7 @@ type (
 		ID               primitive.ObjectID `json:"_id"`
 		Name             string             `json:"name"`
 		CashbackPercent  float64            `json:"cashbackPercent"`
-		TotalTransaction int                `json:"totalTransaction"`
+		TotalTransaction int64              `json:"totalTransaction"`
 		TotalRevenue     float64            `json:"totalRevenue"`
 	}
 )
